feat(config): add application name as part of the log information
This will help when there is a custom domain. Our monitoring solution is going to use this to parse out information

The namespace is prepended to the application name only when the namespace and app name (as known by the service) are not the same. This happens with the deis namespace