fix(model): Specify a default ssl cipher list
This specifies a default cipher list instead of relying on the defaults
used by openssl, which are not tailored to https and cause issues with
HTTP/2 in Firefox due to poor cipher order.

The defaults are chosen to be secure, fast and compatible with current
clients, but exclude IE6 on XP and Java6 from connecting.

The list was built using current recommendation from the Mozilla Wiki
at https://wiki.mozilla.org/Security/Server_Side_TLS for intermediate
compatibility.

Ciphers are preferred in this order:

- Prefer 128-Bit over 256-Bit encryptions (lower coputational overhead)
- Prefer GCM over EDH over RSA auth (GCM & EDH support Forward Secrecy)
- Fallback to 112-Bit 3DES (mainly for IE 8 compatibility)