fix(nginx): Enable builder PROXY PROTOCOL support
If proxy protocol is enabled on an AWS ELB in kubernetes using the
`service.beta.kubernetes.io/aws-load-balancer-proxy-protocol: '*'`
annotation on the load balancer service, it will enable PROXY PROTOCOL
on all ports forwarded by the ELB, which means the builder proxy on port
2222 also needs to use enable proxy_protocol on its listen directive.

There is currently no way to configure PROXY PROTOCOL per port on AWS
using kubenetes service annotation and it doesn't really make sense to
enable proxy protocol only for port 80 and 443 anyways, so this fix uses
the same configuration setting as is used for http and https.

This fixes #255.